refactor: extracted settings icon :recycle:
