+++ src/App.tsx
@@ -45,7 +45,7 @@ import PATHS from "commons/constants/PATHS";
 import NotFoundPage from "./features/404/NotFound.page";
 import useCheckApiConnection from "./commons/hooks/connection/useCheckApiConnection";
 import scanMenuPage from "./features/computer/ScanMenu/ScanMenu.page";
-import {ComputerIconComponent} from "./commons/components/Icons";
+import {ComputerIconComponent, WheelIconComponent} from "./commons/components/Icons";
 
 setupIonicReact();
 
@@ -78,15 +78,7 @@ function CodebarIconComponent() {
     </svg>;
 }
 
-function WheelIconComponent() {
-    return <svg width="40" height="41" viewBox="0 0 40 41" fill="none"
-                xmlns="http://www.w3.org/2000/svg">
-        <rect x="0.5" y="1" width="39" height="39" rx="19.5" fill="#193549" stroke="#CCCCCC"/>
-        <path
-            d="M10.9091 19.5684L12.7727 19.6616C12.8829 18.6419 13.1983 17.6567 13.699 16.7683C14.1997 15.88 14.8749 15.1077 15.6818 14.5005L14.6636 12.8981C14.4182 12.451 14.5455 11.8827 15 11.6218C15.4545 11.3703 15.9909 11.5194 16.2455 11.9665L17.0909 13.662C18.0099 13.2572 18.9997 13.0484 20 13.0484C21.0003 13.0484 21.9901 13.2572 22.9091 13.662L23.7545 11.9665C24.0091 11.5194 24.5455 11.3703 25 11.6218C25.4545 11.8827 25.5818 12.451 25.3364 12.8981L24.3182 14.5005C25.1251 15.1077 25.8003 15.88 26.301 16.7683C26.8017 17.6567 27.1171 18.6419 27.2273 19.6616L29.0909 19.5684C29.332 19.5684 29.5632 19.6665 29.7337 19.8413C29.9042 20.016 30 20.2529 30 20.5C30 20.7471 29.9042 20.984 29.7337 21.1587C29.5632 21.3335 29.332 21.4316 29.0909 21.4316L27.2273 21.3384C27.1171 22.3581 26.8017 23.3433 26.301 24.2317C25.8003 25.12 25.1251 25.8923 24.3182 26.4995L25.3364 28.1019C25.5818 28.549 25.4545 29.1173 25 29.3782C24.5455 29.6297 24.0091 29.4806 23.7545 29.0335L22.9091 27.338C21.9901 27.7428 21.0003 27.9516 20 27.9516C18.9997 27.9516 18.0099 27.7428 17.0909 27.338L16.2455 29.0335C15.9909 29.4806 15.4545 29.6297 15 29.3782C14.5455 29.1173 14.4182 28.549 14.6636 28.1019L15.6818 26.4995C14.8749 25.8923 14.1997 25.12 13.699 24.2317C13.1983 23.3433 12.8829 22.3581 12.7727 21.3384L10.9091 21.4316C10.668 21.4316 10.4368 21.3335 10.2663 21.1587C10.0958 20.984 10 20.7471 10 20.5C10 20.2529 10.0958 20.016 10.2663 19.8413C10.4368 19.6665 10.668 19.5684 10.9091 19.5684ZM17.3364 19.8945C17.4545 19.3262 17.7545 18.8231 18.1818 18.4412L16.6727 16.0749C15.5322 16.9719 14.7838 18.2947 14.5909 19.7547L17.3364 19.8945ZM20 17.7052C20.2909 17.7052 20.5636 17.7518 20.8182 17.8356L22.0727 15.3296C21.4364 15.0594 20.7364 14.9104 20 14.9104C19.2636 14.9104 18.5636 15.0594 17.9273 15.3296L19.1818 17.8356C19.4364 17.7518 19.7091 17.7052 20 17.7052ZM22.6636 19.8945L25.4091 19.7547C25.2162 18.2947 24.4678 16.9719 23.3273 16.0749L21.8182 18.4412C22.2455 18.8231 22.5455 19.3262 22.6636 19.8945ZM22.6636 21.1055C22.5455 21.6738 22.2455 22.1769 21.8182 22.5588L23.3273 24.9251C24.4678 24.028 25.2162 22.7053 25.4091 21.2453L22.6636 21.1055ZM20 23.2948C19.7091 23.2948 19.4364 23.2482 19.1727 23.1644L17.9273 25.6704C18.5636 25.9406 19.2636 26.0896 20 26.0896C20.7364 26.0896 21.4364 25.9406 22.0727 25.6704L20.8273 23.1644C20.5636 23.2482 20.2909 23.2948 20 23.2948ZM17.3364 21.1055L14.5909 21.2453C14.7909 22.7358 15.5545 24.0494 16.6727 24.9251L18.1818 22.5588C17.7545 22.1769 17.4545 21.6738 17.3364 21.1055Z"
-            fill="#CCCCCC"/>
-    </svg>;
-}
+
 
 const withConnectionStatus = (Component: React.FC) => (props: any) => {
