+++ src/commons/components/Icons/index.tsx
@@ -1,4 +1,5 @@
 import AlertIconComponent from "./AlertIcon/AlertIcon.component";
 import ComputerIconComponent from "./ComputerIcon/ComputerIcon.component";
+import WheelIconComponent from "./WheelIcon/WheelIcon.component";
 
-export {AlertIconComponent, ComputerIconComponent};
+export {AlertIconComponent, ComputerIconComponent, WheelIconComponent};
